requirements.txt
pandas
faiss-cpu
saiph
synthcity
synthetic-data-generation
syntheval
numpy==1.24.3
scipy==1.10.0
networkx==2.6.3
cupy==11.6.0
mpmath==1.3.0